+++ src/main/java/com/example/ventas_bodega/dto/interfaces/HistoryStockDtoInter.java
@@ -1,5 +1,6 @@
 package com.example.ventas_bodega.dto.interfaces;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface HistoryStockDtoInter {
@@ -10,11 +11,11 @@ public interface HistoryStockDtoInter {
 
     LocalDateTime getCreateDate();
 
-    Long getStockBefore();
+    BigDecimal getStockBefore();
 
-    Long getStockAfter();
+    BigDecimal getStockAfter();
 
-    Long getStockVariation();
+    BigDecimal getStockVariation();
 
     Long getCompanyId();
